Moved add/remove CosmosDB connection methods to CosmosTreeItem class.
