+++ src/extension.ts
@@ -12,14 +12,14 @@ import { AppSettingsTreeItem, AppSettingTreeItem, editScmType, getFile, IFileRes
 import { AzureTreeDataProvider, AzureUserInput, IActionContext, IAzureNode, IAzureParentNode, IAzureTreeItem, IAzureUserInput, registerCommand, registerEvent, registerUIExtensionVariables } from 'vscode-azureextensionui';
 import TelemetryReporter from 'vscode-extension-telemetry';
 import { SiteConfigResource } from '../node_modules/azure-arm-website/lib/models';
-import { addCosmosDBConnection } from './commands/connections/addCosmosDBConnection';
-import { removeCosmosDBConnection } from './commands/connections/removeCosmosDBConnection';
 import { deploy } from './commands/deploy';
 import { enableFileLogging } from './commands/enableFileLogging';
 import { disableRemoteDebug } from './commands/remoteDebug/disableRemoteDebug';
 import { startRemoteDebug } from './commands/remoteDebug/startRemoteDebug';
 import { startStreamingLogs } from './commands/startStreamingLogs';
 import { swapSlots } from './commands/swapSlots';
+import { CosmosDBDatabase } from './explorer/CosmosDBDatabase';
+import { CosmosDBTreeItem } from './explorer/CosmosDBTreeItem';
 import { DeploymentSlotsNATreeItem, DeploymentSlotsTreeItem, ScaleUpTreeItem } from './explorer/DeploymentSlotsTreeItem';
 import { DeploymentSlotTreeItem } from './explorer/DeploymentSlotTreeItem';
 import { FileEditor } from './explorer/editors/FileEditor';
@@ -324,8 +324,22 @@ export function activate(context: vscode.ExtensionContext): void {
             opn('https://marketplace.visualstudio.com/items?itemName=ms-azuretools.vscode-cosmosdb');
         }
     });
-    registerCommand('appService.AddCosmosDBConnection', addCosmosDBConnection);
-    registerCommand('appService.RemoveCosmosDBConnection', removeCosmosDBConnection);
+    registerCommand('appService.AddCosmosDBConnection', async (node: IAzureParentNode<CosmosDBTreeItem>) => {
+        if (node.treeItem.contextValue === 'AddCosmosDBConnection') {
+            // tslint:disable-next-line:no-non-null-assertion
+            const parentNode = node.parent!;
+            await parentNode.createChild();
+            parentNode.refresh();
+        } else {
+            await node.createChild();
+        }
+    });
+    registerCommand('appService.RemoveCosmosDBConnection', async (node: IAzureNode<CosmosDBDatabase>) => {
+        // tslint:disable-next-line:no-non-null-assertion
+        const parentNode = <IAzureParentNode<CosmosDBTreeItem>>node.parent!;
+        parentNode.treeItem.deleteTreeItem(node);
+        parentNode.refresh();
+    });
 }
 
 // tslint:disable-next-line:no-empty
